Adding more cruds

+++ src/main/java/pe/com/aldesa/aduanero/repository/CamionRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import pe.com.aldesa.aduanero.entity.Camion;
 
 @Repository
-public interface CamionRepository extends JpaRepository<Camion, Integer> {
+public interface CamionRepository extends JpaRepository<Camion, Long> {
 
 }
